+++ src/components/10-atoms/input-text/ui.test.js
@@ -210,11 +210,11 @@ test('should display correct default value', async t => {
   await t.expect(await getValue()).eql('qwertz');
 });
 
-fixture('Input-Text React OnInputfieldKeyUp').page(
-  `${host}/iframe.html?id=examples-datepicker-react--using-oninputfieldkeyup-event`
+fixture('Input-Text React OnInput').page(
+  `${host}/iframe.html?id=examples-datepicker-react--using-oninput-event`
 );
 
-test('should fire onInputfieldKeyUp callback on user input', async t => {
+test('should fire onInput callback on user input', async t => {
   const testoutput = await Selector(() =>
     document.querySelector('#inputtext-react-testoutput')
   ).addCustomDOMProperties({